Run the linter on the Vue directory, and perform manual fixes.
Most of the manual fixes were vbind:key resommendations. I also had to rename
Baduk.vue to a two-word name. BadukBoard seemed reasonable, but then it also
made sense to change the directory name from variants to boards.
