+++ packages/vue-client/src/components/GameList.vue
@@ -1,17 +1,16 @@
 <script setup lang="ts">
-  import { storeToRefs } from "pinia";
-  import { useStore } from "../stores/games";
-  import router from "../router";
+import { storeToRefs } from "pinia";
+import { useStore } from "../stores/games";
 
-  const store = useStore();
-  const { games } = storeToRefs(store);
+const store = useStore();
+const { games } = storeToRefs(store);
 
-  store.fetchGames();
+store.fetchGames();
 </script>
 
 <template>
   <ul>
-    <li v-for="game in games">
+    <li v-for="game in games" :key="game.id">
       <RouterLink v-bind:to="{ name: 'game', params: { gameId: game.id } }">
         {{ game.id }}
       </RouterLink>
