+++ packages/vue-client/src/components/TheWelcome.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
-import WelcomeItem from './WelcomeItem.vue'
-import DocumentationIcon from './icons/IconDocumentation.vue'
-import ToolingIcon from './icons/IconTooling.vue'
-import EcosystemIcon from './icons/IconEcosystem.vue'
-import CommunityIcon from './icons/IconCommunity.vue'
-import SupportIcon from './icons/IconSupport.vue'
+import WelcomeItem from "./WelcomeItem.vue";
+import DocumentationIcon from "./icons/IconDocumentation.vue";
+import ToolingIcon from "./icons/IconTooling.vue";
+import EcosystemIcon from "./icons/IconEcosystem.vue";
+import CommunityIcon from "./icons/IconCommunity.vue";
+import SupportIcon from "./icons/IconSupport.vue";
 </script>
 
 <template>
@@ -26,10 +26,11 @@ import SupportIcon from './icons/IconSupport.vue'
     <template #heading>Tooling</template>
 
     This project is served and bundled with
-    <a href="https://vitejs.dev/guide/features.html" target="_blank">Vite</a>. The recommended IDE
-    setup is <a href="https://code.visualstudio.com/" target="_blank">VSCode</a> +
-    <a href="https://github.com/johnsoncodehk/volar" target="_blank">Volar</a>. If you need to test
-    your components and web pages, check out
+    <a href="https://vitejs.dev/guide/features.html" target="_blank">Vite</a>.
+    The recommended IDE setup is
+    <a href="https://code.visualstudio.com/" target="_blank">VSCode</a> +
+    <a href="https://github.com/johnsoncodehk/volar" target="_blank">Volar</a>.
+    If you need to test your components and web pages, check out
     <a href="https://www.cypress.io/" target="_blank">Cypress</a> and
     <a href="https://on.cypress.io/component" target="_blank"
       >Cypress Component Testing</a
@@ -49,10 +50,13 @@ import SupportIcon from './icons/IconSupport.vue'
     Get official tools and libraries for your project:
     <a target="_blank" href="https://pinia.vuejs.org/">Pinia</a>,
     <a target="_blank" href="https://router.vuejs.org/">Vue Router</a>,
-    <a target="_blank" href="https://test-utils.vuejs.org/">Vue Test Utils</a>, and
-    <a target="_blank" href="https://github.com/vuejs/devtools">Vue Dev Tools</a>. If you need more
-    resources, we suggest paying
-    <a target="_blank" href="https://github.com/vuejs/awesome-vue">Awesome Vue</a>
+    <a target="_blank" href="https://test-utils.vuejs.org/">Vue Test Utils</a>,
+    and
+    <a target="_blank" href="https://github.com/vuejs/devtools">Vue Dev Tools</a
+    >. If you need more resources, we suggest paying
+    <a target="_blank" href="https://github.com/vuejs/awesome-vue"
+      >Awesome Vue</a
+    >
     a visit.
   </WelcomeItem>
 
@@ -63,10 +67,13 @@ import SupportIcon from './icons/IconSupport.vue'
     <template #heading>Community</template>
 
     Got stuck? Ask your question on
-    <a target="_blank" href="https://chat.vuejs.org">Vue Land</a>, our official Discord server, or
-    <a target="_blank" href="https://stackoverflow.com/questions/tagged/vue.js">StackOverflow</a>.
-    You should also subscribe to
-    <a target="_blank" href="https://news.vuejs.org">our mailing list</a> and follow the official
+    <a target="_blank" href="https://chat.vuejs.org">Vue Land</a>, our official
+    Discord server, or
+    <a target="_blank" href="https://stackoverflow.com/questions/tagged/vue.js"
+      >StackOverflow</a
+    >. You should also subscribe to
+    <a target="_blank" href="https://news.vuejs.org">our mailing list</a> and
+    follow the official
     <a target="_blank" href="https://twitter.com/vuejs">@vuejs</a>
     twitter account for latest news in the Vue world.
   </WelcomeItem>
@@ -77,8 +84,8 @@ import SupportIcon from './icons/IconSupport.vue'
     </template>
     <template #heading>Support Vue</template>
 
-    As an independent project, Vue relies on community backing for its sustainability. You can help
-    us by
+    As an independent project, Vue relies on community backing for its
+    sustainability. You can help us by
     <a target="_blank" href="https://vuejs.org/sponsor/">becoming a sponsor</a>.
   </WelcomeItem>
 </template>
