+++ packages/vue-client/src/stores/counter.ts
@@ -1,16 +1,16 @@
-import { defineStore } from 'pinia'
+import { defineStore } from "pinia";
 
 export const useCounterStore = defineStore({
-  id: 'counter',
+  id: "counter",
   state: () => ({
-    counter: 0
+    counter: 0,
   }),
   getters: {
-    doubleCount: (state) => state.counter * 2
+    doubleCount: (state) => state.counter * 2,
   },
   actions: {
     increment() {
-      this.counter++
-    }
-  }
-})
+      this.counter++;
+    },
+  },
+});
