+++ packages/vue-client/src/views/AboutView.vue
@@ -17,11 +17,11 @@
 </template>
 
 <style>
-  @media (min-width: 1024px) {
-    .about {
-      min-height: 100vh;
-      display: flex;
-      align-items: center;
-    }
+@media (min-width: 1024px) {
+  .about {
+    min-height: 100vh;
+    display: flex;
+    align-items: center;
   }
+}
 </style>
